+++ WebApi/Interfaces/ILocationPointRepository.cs
@@ -5,7 +5,7 @@ namespace WebApi.Interfaces
     public interface ILocationPointRepository
     {
         void AddLocationPoint(LocationPoint locationPoint);
-        Task<LocationPoint> GetLocationPointAsync(int id);
+        Task<LocationPoint> GetLocationPointAsync(long id);
         Task<bool> CheckCordinatesAsync(double latitude, double longitude);
         void UpdateLocationPoint(LocationPoint locationPoint);
         void DeleteLocationPoint(LocationPoint locationPoint);
